Added activity class
An activity is a atomic element that can executed in a job. It takes an
input and is aware of the imutable context in which it is executed in
order to provide an output.

The input can be any combination of any previous activity's output.
An activity defines some or no transitions which will be executed after
it.