Added transform library
The transform library is used to transform a json/js object into another.
It takes 2 parameters (the mapping, and the target) and produces an output
that matches the mapping keys.

However if a key is _jsonPath, the value of that key will be the result of
the json path specified as value of _jsonPath applied to the target.

If a key on top of having a _jsonPath has a _transformation property, it
will apply the transformation method to the result of the previously
obtained value and assing its result to the output's matching key.