Added JMSListen activity
The JSMListen activity is a generator activity. It means that until
stopped it will keep on running. It takes a connection object containing a
host and a port along with a destination queue as parameters.

It will listen to the given queue on the given connection and emit any
message that comes in.

Messages will not be automaticaly acknowledged. Using the JMSAcknowledge
is mandatory in order to dequeue the message.