Added iteration activity.
The iteration activity will iterate over the input array and execute a
chain of activities for each item. The chain starts with a starter and
ends with a finisher. The finisher is only used to generate results, it
does not need to be the last element of the chain.
The finisher's of the last execution of the chain will be passed on to the
next activity.

If the 'aggregateResult' boolean is set to true in the
parameters. The iterate activity will return an array of all the results
outputted by the finisher.