Updated job runner to expose 'chainRun' method
The chain run method was modified in 2 ways.

1- It is now in the Runner class as a static method to allow it being used
by other modules. This change was needed because some of the orchestration
components (e.g. the iterate activity) might need to run some activities.

2- It now calls a callback function (with no arguments) when the whole
chain is executed. This helps with orchestration as we can execute a list
of activities after an other or simply start an action once a list of
activities is done.