Added postgres connect activity
This activity will take a connection (username, host, port, password,
database) and establish a connection. It also requires an id parameter
which will become the connection id and can be reused in further
activities.

On top of establishing the connection, to ensure good practice it will
start a transaction (effectively sending a 'BEGIN' command to postgres).
To validate and commit that transaction please use the PostgresCommit
Activity.

Once established you can use any number of PostgresQuery activities before
commiting or rolling back. To roll back, you can use the PostgresRollback
components.

PostgresCommit, PostgresRollback, PostgresQuery will be added within the
next commits.

If there is already a connection registered on the id it will not recreate
it but simply start a new transaction.