Added PostgresQuery activity
It takes as parameters a connection (an object with username, host, port, password,
database or a string with a connection id), a query string that must be a
valid parametrised postgres query along with an args parameter that contain the array
of arguments to be givent to the parametrised query.

To learn more on what queries are acceptable read the postgres
documentation for you version of postgres
(http://www.postgresql.org/docs/).

The underlying implementation is based on the node-postgres module
(https://github.com/brianc/node-postgres) and is fully compatible with any
of its features.

2 Additions have been made to that module for this activity.

1- Given an array of queries rather than a string it will sequentially
call all the queries. These queries might be parametrised with a
multidimensional args array.

2- Whenever there is a request requiring an 'IN' you can write it as
follows:
"SELECT name FROM users WHERE privileges = $1 and id IN $2;"
And give the following args array:
['admin',[5,2,7,20,14,4,11,1]]

It will generate this query:
"SELECT name FROM users WHERE privileges = 'admin' and id IN (5,2,7,20,14,4,11,1);"